Exercise 1.1 refactoring out components
Here I refactored the root to render 3 separate components.
One for each of the header, content and total.
Ideally, I'd like to make these components more generic,
as Total, although having a generic
title, is very specifically summing the
value of exercises and printing a hard-coded message
regarding those exercises. It may be better to just have a component
that takes a "message" prop and three
"value" props instead. I will continue to follow the course however and
make this change if it makes sense to do so.