Exercise 2.12 data for countries step 1
Created an application that retrieves data on countries from
a public REST API endpoint and allows you to filter out and display
detailed information on a country.

I was not required to satisfy the edge case where the name
of the country is part of the name of another country for this
exercise.