Exercise 2.13 data for countries step2
Added show buttons beside each country when a list of countries is
displayed that shows the detailed view for that country when pressed.
Used a piece of state called "detailedCountry" that is set by
pressing the show button. When it is set, the detailed view will show.
I kept all pre-existing functionality intact. Whenever a new character
is added to the filter, it clears the detailed view so that if you
narrow the filter to one country, it will still display that country'set
detailed view.